pyproject.toml
[project]
name = "shelf"
version = "1.0.0"
description = "self hosted markdown-backed shelf"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "python-frontmatter>=1.1.0",
    "python-slugify>=8.0.0",
    "ruff>=0.14.10",
]
